+++ CHANGELOG.md
@@ -20,8 +20,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Adds the mobile navigation menu for the `Header` component with `MobileNav` and `MobileNavButton` components.
 - Adds the mobile login menu for the `Header` component.
 - Adds the `useCloseDropDown` hook to manage closing dropdown-like components through clicks outside of it and by pressing the escape key.
-- Adds React context for the `Header`'s patron state data through `PatronContext` and `PatronProvider`.
-- Fixes `useNYPLBreakpoints` hook so it internally controls its own state.
+- Adds React context for the `Header`'s patron state data through `HeaderPatronContext` and `HeaderPatronProvider`.
+- Better namespacing for Header-only components.
 
 ## Prerelease
