+++ src/components/Header/Header.test.tsx
@@ -6,14 +6,14 @@ import { axe } from "jest-axe";
 import renderer from "react-test-renderer";
 
 import Header from "./Header";
-import { alertsApiUrl, tokenRefreshLink } from "./utils/headerUtils";
 import {
   patronApiUrlWithToken,
   mockExpiredResponseData,
   mockLoginCookie,
   mockResponseData,
 } from "./utils/authApiMockResponse";
-import { refineryResponse } from "./utils/SitewideAlertsMocks";
+import { alertsApiUrl, tokenRefreshLink } from "./utils/headerUtils";
+import { refineryResponse } from "./utils/sitewideAlertsMocks";
 
 // We want to mock the media queries and set the desktop view.
 jest.mock("../../hooks/useNYPLBreakpoints", () => {
