+++ src/components/Header/components/HeaderLogin.test.tsx
@@ -2,24 +2,24 @@ import { render, screen, within } from "@testing-library/react";
 import { axe } from "jest-axe";
 import * as React from "react";
 
-import Login from "./Login";
+import HeaderLogin from "./HeaderLogin";
 
-describe("Login Accessibility", () => {
+describe("HeaderLogin Accessibility", () => {
   it("passes axe accessibility test", async () => {
-    const { container } = render(<Login patronName="" />);
+    const { container } = render(<HeaderLogin patronName="" />);
     expect(await axe(container)).toHaveNoViolations();
   });
 
   it("passes axe accessibility test for mobile", async () => {
-    const { container } = render(<Login isMobile patronName="" />);
+    const { container } = render(<HeaderLogin isMobile patronName="" />);
     expect(await axe(container)).toHaveNoViolations();
   });
 });
 
-describe("Login", () => {
+describe("HeaderLogin", () => {
   describe("Desktop", () => {
     it("renders the logged out UI if there is no `patronName` value", () => {
-      render(<Login patronName="" />);
+      render(<HeaderLogin patronName="" />);
 
       const links = screen.getAllByRole("link");
 
@@ -29,7 +29,7 @@ describe("Login", () => {
     });
 
     it("renders the logged in UI if there is a `patronName` value", () => {
-      render(<Login patronName="PATRON, JANE A" />);
+      render(<HeaderLogin patronName="PATRON, JANE A" />);
 
       const greetingContainer = screen.getByTestId("patronGreeting");
       const links = screen.getAllByRole("link");
@@ -50,7 +50,7 @@ describe("Login", () => {
 
   describe("Mobile", () => {
     it("renders the logged out UI if there is no `patronName` value", () => {
-      render(<Login isMobile patronName="" />);
+      render(<HeaderLogin isMobile patronName="" />);
 
       const links = screen.getAllByRole("link");
 
@@ -60,7 +60,7 @@ describe("Login", () => {
     });
 
     it("renders the logged in UI if there is a `patronName` value", () => {
-      render(<Login isMobile patronName="PATRON, JANE A" />);
+      render(<HeaderLogin isMobile patronName="PATRON, JANE A" />);
 
       const greetingContainer = screen.getByTestId("patronGreeting");
       const links = screen.getAllByRole("link");
