+++ src/components/Header/components/HeaderMobileIconNav.tsx
@@ -3,26 +3,26 @@ import { chakra, Flex, useStyleConfig } from "@chakra-ui/react";
 
 import Icon from "../../Icons/Icon";
 import Link from "../../Link/Link";
-import LoginButton from "./LoginButton";
-import MobileNavButton from "./MobileNavButton";
-import SearchButton from "./SearchButton";
+import HeaderLoginButton from "./HeaderLoginButton";
+import HeaderMobileNavButton from "./HeaderMobileNavButton";
+import HeaderSearchButton from "./HeaderSearchButton";
 
-const MobileIconNav = chakra(() => {
+const HeaderMobileIconNav = chakra(() => {
   const styles = useStyleConfig("HeaderMobileIconNav");
 
   return (
     <Flex sx={styles}>
-      <LoginButton isMobile />
+      <HeaderLoginButton isMobile />
       <Link
         aria-label="NYPL Locations Near Me"
         href="https://nypl.org/locations"
       >
         <Icon name="locator" size="large" title="NYPL Locator" />
       </Link>
-      <SearchButton isMobile />
-      <MobileNavButton />
+      <HeaderSearchButton isMobile />
+      <HeaderMobileNavButton />
     </Flex>
   );
 });
 
-export default MobileIconNav;
+export default HeaderMobileIconNav;
