--- src/components/Header/components/index.tsx
@@ -1,27 +1,0 @@
-// Header-only components
-import Login from "./Login";
-import LoginButton from "./LoginButton";
-import LowerNav from "./LowerNav";
-import MobileIconNav from "./MobileIconNav";
-import MobileNav from "./MobileNav";
-import MobileNavButton from "./MobileNavButton";
-import SearchButton from "./SearchButton";
-import SearchForm from "./SearchForm";
-import SitewideAlerts from "./SitewideAlerts";
-import UpperNav from "./UpperNav";
-// React Context
-import { PatronProvider } from "../context/patronContext";
-
-export default {
-  Login,
-  LoginButton,
-  LowerNav,
-  MobileIconNav,
-  MobileNav,
-  MobileNavButton,
-  PatronProvider,
-  SearchButton,
-  SearchForm,
-  SitewideAlerts,
-  UpperNav,
-};
